pyproject.toml
[tool.poetry]
name = "optimizer"
version = "0.1.0"
description = ""
authors = ["Dancing Coconut <dancingcoconut12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.5.2"
eco2ai = "^0.3.9"
streamlit = "^1.33.0"
pandas = "^1.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
streamlit = "app.main"